Move db test fixture to conftest
